Emergency HOTFIX - Fixed Login hash bug
As you can probably see, the culprit was db.any. If you're curious, the reason this caused an error is because this led to queryResults being a JSON object with an array in it, with that array only having one object. Thus, to access the hashed password returned from that array, you would have to call queryResults[0].password. But our code was written to call queryResults.password. There were 2 ways to fix this: fix the db.any call, or fix the queryResult accessing. I opted to change the db.any call to db.one, and this way the JSON object is not an array, and its elements can be accessed without the brackets and index. Cheers, everyone and good luck with your work